\name{cor_afe}
\alias{VenturaleonTools}
\title{Herramientos del profesor Jose Ventura Leon}

\description{
This package provides a function to examine the correlations between the factors.
}

\usage{
cor_afe(factors_data)
}

\arguments{
\item{factors_data}{It is the object of the psych library}


}


\examples{
cor_afe(factors_data)
}
