\name{crear_modelo_lavaan}
\alias{crear_modelo_lavaan}
\title{Tools of Professor Jose Ventura Leon}

\description{
This function allows you to generate a lavaan model quickly.
}

\usage{
crear_modelo_lavaan(nombre)
}

\arguments{
\item{nombre}{These are the names of the items or indicators}
}


\examples{
crear_modelo_lavaan(nombre = "CCOV", F1 = c(3, 6, 2))
}
