\name{diag_aba_na}
\alias{diag_aba_na}
\title{Tools of Professor Jose Ventura Leon}

\description{
This package provides a function to obtain the matrix only from the diagonal down.
}

\usage{
diag_aba_na(matriz)
}

\arguments{
\item{matriz}{It is a data matrix}
}


\examples{
diag_aba_na(matriz)
}
