\name{extract_fit}
\alias{VenturaleonTools}
\title{Herramientos del profesor Jose Ventura Leon}

\description{
This package offers a function to extract the goodness of fit from a psych object.
}

\usage{
extract_fit(factors_data)
}

\arguments{
\item{factors_data}{It is the object of the psych library}
}


\examples{
extract_fit(factors_data)
}
