\name{extract_fit_measures}
\alias{VenturaleonTools}
\title{Tools of Professor Jose Ventura Leon}

\description{
This package provides a function to extract goodness of fit from exploratory factor models with lavaan.
}

\usage{
extract_fit_measures(Specifications)
}

\arguments{
\item{Specifications}{It is the model obtained with the specification_models function.}

}


\examples{
extract_fit_measures(Specifications)
}
