\name{extracted_items2}
\alias{extracted_items2}
\title{Tools of Professor Jose Ventura Leon}

\description{
This function extracts the items of each of the factors to be used in the calculation of the reliability of the calcula_omega_all function.
}

\usage{
extracted_items2(data)
}

\arguments{
\item{data}{is the dataframe containing the factor loadings of the AFE model.}
}


\examples{
extracted_items2(data)
}
