\name{ExtractItems}
\alias{VenturaleonTools}
\title{Tools of Professor Jose Ventura Leon}

\description{
This package offers a function to extract the name of the items according to each of the test factors.
}

\usage{
extract_items(text)
}

\arguments{
\item{text}{The text where the factors are found with each of their respective items}
}


\examples{
extracted$[variable name]
}
