\name{grafico_boxplots}
\alias{grafico_boxplots}
\title{Tools of Professor Jose Ventura Leon}

\description{
This package provides a function for a quick box plot of all the variables under study.
}

\usage{
grafico_boxplots(Data, columnas)
}

\arguments{
\item{Data}{It is the database}
\item{columnas}{It is an object containing the name of the variables under study.}
}

\examples{
grafico_boxplots(Data, columnas)
}
