\name{Multivariate_plot}
\alias{Multivariate_plot}
\title{Tools of Professor Jose Ventura Leon}

\description{
This function provides a plot of QQ-Plot and Mardia coefficient at the same time.
}

\usage{
Multivariate_plot(data, xmin=30, xmax=40, ymin=2, ymax=7)
}

\arguments{
\item{data}{It is the database}
\item{xmin}{En caso desee mover el grafico en el eje x para que comience en el valor establecido}
\item{xmax}{En caso desee mover el grafico en el eje x para que finalice en el valor establecido}
\item{ymin}{En caso desee mover el grafico en el eje y para que comience en el valor establecido}
\item{ymax}{En caso desee mover el grafico en el eje y para que finalice en el valor establecido}

}

\examples{
Multivariate_plot(data)
}
